Start of remove package deps refactor
