+++ weave/decorator_class.py
@@ -2,7 +2,8 @@ import inspect
 
 from . import registry_mem
 from . import op_def
-from . import derive_op
+
+# from . import derive_op
 from . import errors
 from . import weave_types as types
 
@@ -23,7 +24,7 @@ def weave_class(weave_type: type[types.Type]):
                     opdef.input_type.arg_types["self"] = weave_type()
                     # Now that we have a self_type, we may be able to derive
                     # ops.
-                    derive_op.derive_ops(opdef)
+                    # derive_op.derive_ops(opdef)
                 # Replace function op names with method op names
                 current_name = opdef.name
                 requires_rename = current_name.startswith("op-")
@@ -38,8 +39,9 @@ def weave_class(weave_type: type[types.Type]):
                     )
 
                     for derived_handler_id, op in member.derived_ops.items():
-                        handler = derive_op.handler_for_id(derived_handler_id)
-                        handler.handle_class_decorator_update(op, weave_type, new_name)
+                        # handler = derive_op.handler_for_id(derived_handler_id)
+                        # handler.handle_class_decorator_update(op, weave_type, new_name)
+                        pass
 
         # Check __dict__ instead of using regular attribute access
         # because we want to add instance_classes even if it is already
