+++ weave/mappers_publisher.py
@@ -12,7 +12,8 @@ from . import mappers_python_def
 from .language_features.tagging import tagged_value_type
 import dataclasses
 from weave import weave_internal, context, storage
-from .ops_primitives import weave_api
+
+# from .ops_primitives import weave_api
 from . import artifact_local
