+++ weave/registry_mem.py
@@ -3,8 +3,6 @@ import datetime
 
 from weave.op_args import OpNamedArgs
 
-from . import op_def
-from . import op_def_type
 from . import op_args
 from . import weave_types
 from . import errors
@@ -13,6 +11,9 @@ from . import storage
 from . import uris
 from . import op_aliases
 
+if typing.TYPE_CHECKING:
+    from .op_def import OpDef
+
 
 class Registry:
     _types: typing.Dict[str, weave_types.Type]
@@ -20,13 +21,13 @@ class Registry:
     # This most recent register_op() call for a given OpDef.name
     # TODO: Get rid of this! Always use versioning! This is a temporary
     # state.
-    _ops: typing.Dict[str, op_def.OpDef]
+    _ops: typing.Dict[str, "OpDef"]
 
     # common_name: name: op_def
-    _ops_by_common_name: typing.Dict[str, dict[str, op_def.OpDef]]
+    _ops_by_common_name: typing.Dict[str, dict[str, "OpDef"]]
 
     # Ops stored by their URI (for ops that are non-builtin).
-    _op_versions: typing.Dict[str, op_def.OpDef]
+    _op_versions: typing.Dict[str, "OpDef"]
 
     # Maintains a timestamp of when the registry was last updated.
     # This is useful for caching the ops dictionary when serving
@@ -46,7 +47,7 @@ class Registry:
     def updated_at(self) -> float:
         return self._updated_at
 
-    def register_op(self, op: op_def.OpDef, location=None):
+    def register_op(self, op: "OpDef", location=None):
         if context_state.get_no_op_register():
             return op
         self.mark_updated()
@@ -77,7 +78,7 @@ class Registry:
     def have_op(self, op_name: str) -> bool:
         return op_name in self._ops
 
-    def get_op(self, uri: str) -> op_def.OpDef:
+    def get_op(self, uri: str) -> "OpDef":
         object_uri = uris.WeaveURI.parse(uri)
         if object_uri.version is not None:
             object_key = str(object_uri)
@@ -102,16 +103,14 @@ class Registry:
                 return op_def
         raise Exception("Op def doesn't exist for %s" % lazy_local_fn)
 
-    def find_ops_by_common_name(self, common_name: str) -> typing.List[op_def.OpDef]:
+    def find_ops_by_common_name(self, common_name: str) -> typing.List["OpDef"]:
         aliases = op_aliases.get_op_aliases(common_name)
-        ops: list[op_def.OpDef] = []
+        ops: list["OpDef"] = []
         for alias in aliases:
             ops.extend(self._ops_by_common_name.get(alias, {}).values())
         return ops
 
-    def find_chainable_ops(
-        self, arg0_type: weave_types.Type
-    ) -> typing.List[op_def.OpDef]:
+    def find_chainable_ops(self, arg0_type: weave_types.Type) -> typing.List["OpDef"]:
         def is_chainable(op):
             if not isinstance(op.input_type, op_args.OpNamedArgs):
                 return False
@@ -123,6 +122,8 @@ class Registry:
         return [op for op in self._ops.values() if is_chainable(op)]
 
     def load_saved_ops(self):
+        from . import op_def_type
+
         for op_ref in storage.objects(op_def_type.OpDefType()):
             try:
                 op_ref.get()
@@ -130,7 +131,7 @@ class Registry:
                 # print("Failed to load non-builtin op: %s" % op_ref.uri)
                 pass
 
-    def list_ops(self) -> typing.List[op_def.OpDef]:
+    def list_ops(self) -> typing.List["OpDef"]:
         # Note this uses self._ops, so provides the most recent registered op, which could
         # be the last one we loaded() [rather than the last one the user declared] which
         # is incorrect behavior
@@ -139,7 +140,7 @@ class Registry:
     # Currently this just returns all ops that take no arguments.
     # Perhaps a better extension is to require a return type that
     # subclasses some abstract package type?
-    def list_packages(self) -> typing.List[op_def.OpDef]:
+    def list_packages(self) -> typing.List["OpDef"]:
         packages = [
             a
             for a in list(self._ops.values())
