+++ weave/util.py
@@ -3,7 +3,8 @@ import random
 import socket
 import string
 import gc, inspect
-import ipynbname
+
+# import ipynbname
 import typing
 from .errors import WeaveFingerprintErrorMixin
 
@@ -52,8 +53,8 @@ def capture_exception_with_sentry_if_available(
     return None
 
 
-def get_notebook_name():
-    return ipynbname.name()
+# def get_notebook_name():
+#     return ipynbname.name()
 
 
 def get_hostname():
